move search input to the separate component
