+++ src/components/search-input/SearchInput.component.tsx
@@ -1,0 +1,40 @@
+import React, { useState } from 'react';
+
+import { SearchIconComponent } from '../icon/SearchIcon.component.tsx';
+
+import styles from './search-input.module.css';
+
+interface SearchInputProps {
+    search?: string;
+    onSearchSubmit: (search: string) => void;
+}
+
+export const SearchInputComponent: React.FC<SearchInputProps> = ({ search, onSearchSubmit }) => {
+    const [searchTitle, setSearchTitle] = useState<string>(search || '');
+
+    const handleSearchSubmit = (event: React.FormEvent) => {
+        event.preventDefault();
+        onSearchSubmit(searchTitle);
+    };
+
+    const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        const target = event.target as HTMLInputElement;
+
+        setSearchTitle(target.value);
+    };
+
+    return (
+        <form className={styles.productsFilterSearchForm} onSubmit={handleSearchSubmit}>
+            <input
+                className={styles.productsFilterSearchInput}
+                type="text"
+                placeholder="Search..."
+                value={searchTitle}
+                onChange={handleSearchChange}
+            />
+            <button className={styles.productsFilterSearchButton} type="submit">
+                <SearchIconComponent title="Search" />
+            </button>
+        </form>
+    );
+};
