+++ src/components/search-input/search-input.module.css
@@ -1,0 +1,44 @@
+:global :root.dark {
+  --products-filter-search-background-color: var(--secondary-grey-color);
+  --products-filter-search-text-color: var(--main-white-color);
+}
+
+:global :root.light {
+  --products-filter-search-background-color: var(--fifth-grey-color);
+  --products-filter-search-text-color: var(--secondary-grey-color);
+}
+
+.productsFilterSearchForm {
+  display: flex;
+  position: relative;
+}
+
+.productsFilterSearchInput {
+  border: 1px solid var(--products-filter-search-background-color);
+  background-color: var(--products-filter-search-background-color);
+  border-radius: 20px;
+  width: 100%;
+  color: var(--products-filter-search-text-color);
+  padding: 11px 50px 11px 20px;
+  outline: none;
+  font-family: inherit;
+}
+
+.productsFilterSearchInput::placeholder {
+  color: var(--products-filter-search-text-color);
+  font-size: 14px;
+}
+
+.productsFilterSearchButton {
+  position: absolute;
+  top: 0;
+  right: 0;
+  width: 40px;
+  height: 40px;
+  background-color: var(--main-red-color);
+  border: none;
+  border-radius: 40px;
+  padding: 0;
+  cursor: pointer;
+  outline: none;
+}
